+++ packages/cloud-cognitive/src/components/InlineEdit/InlineEdit.js
@@ -73,7 +73,7 @@ export let InlineEdit = React.forwardRef(
     refIn
   ) => {
     const carbonPrefix = usePrefix();
-    const refInput = useRef(null);
+    const refInput = useRef({ innerText: value });
     const localRef = useRef(null);
     const ref = refIn || localRef;
     const [editing, setEditing] = useState(false);
@@ -291,7 +291,7 @@ export let InlineEdit = React.forwardRef(
           onPaste={handlePaste}
           suppressContentEditableWarning={true}
           ref={refInput}
-          data-placeholder={placeholder ?? labelText}
+          data-placeholder={placeholder}
         >
           {value}
         </div>